+++ internal/enginenetx/network_test.go
@@ -2,7 +2,6 @@ package enginenetx_test
 
 import (
 	"context"
-	"encoding/json"
 	"net"
 	"net/http"
 	"net/url"
@@ -10,15 +9,12 @@ import (
 	"time"
 
 	"github.com/apex/log"
-	"github.com/google/go-cmp/cmp"
 	"github.com/ooni/probe-cli/v3/internal/bytecounter"
 	"github.com/ooni/probe-cli/v3/internal/enginenetx"
 	"github.com/ooni/probe-cli/v3/internal/kvstore"
 	"github.com/ooni/probe-cli/v3/internal/measurexlite"
-	"github.com/ooni/probe-cli/v3/internal/model"
 	"github.com/ooni/probe-cli/v3/internal/netemx"
 	"github.com/ooni/probe-cli/v3/internal/netxlite"
-	"github.com/ooni/probe-cli/v3/internal/runtimex"
 	"github.com/ooni/probe-cli/v3/internal/testingsocks5"
 	"github.com/ooni/probe-cli/v3/internal/testingx"
 )
@@ -271,90 +267,6 @@ func TestNetworkQA(t *testing.T) {
 		client := txp.NewHTTPClient()
 		if client.Jar == nil {
 			t.Fatal("expected non-nil cookie jar")
-		}
-	})
-
-	t.Run("NewNetwork uses the correct HTTPSDialerPolicy", func(t *testing.T) {
-		// testcase is a test case run by this func
-		type testcase struct {
-			name         string
-			kvStore      func() model.KeyValueStore
-			expectStatus int
-			expectBody   []byte
-		}
-
-		cases := []testcase{
-			// Without a policy accessing www.example.com should lead to 200 as status
-			// code and the expected web page when we're using netem
-			{
-				name: "when there is no user-provided policy",
-				kvStore: func() model.KeyValueStore {
-					return &kvstore.Memory{}
-				},
-				expectStatus: 200,
-				expectBody:   []byte(netemx.ExampleWebPage),
-			},
-
-			// But we can create a policy that can land us on a different website (not the
-			// typical use case of the policy, but definitely demonstrating it works)
-			{
-				name: "when there's a user-provided policy",
-				kvStore: func() model.KeyValueStore {
-					policy := &enginenetx.HTTPSDialerStaticPolicyRoot{
-						DomainEndpoints: map[string][]*enginenetx.HTTPSDialerTactic{
-							"www.example.com:443": {{
-								Address:        netemx.AddressApiOONIIo,
-								InitialDelay:   0,
-								Port:           "443",
-								SNI:            "www.example.com",
-								VerifyHostname: "api.ooni.io",
-							}},
-						},
-						Version: enginenetx.HTTPSDialerStaticPolicyVersion,
-					}
-					rawPolicy := runtimex.Try1(json.Marshal(policy))
-					kvStore := &kvstore.Memory{}
-					runtimex.Try0(kvStore.Set(enginenetx.HTTPSDialerStaticPolicyKey, rawPolicy))
-					return kvStore
-				},
-				expectStatus: 404,
-				expectBody:   []byte{},
-			},
-		}
-
-		for _, tc := range cases {
-			t.Run(tc.name, func(t *testing.T) {
-				env := netemx.MustNewScenario(netemx.InternetScenario)
-				defer env.Close()
-
-				env.Do(func() {
-					netx := enginenetx.NewNetwork(
-						bytecounter.New(),
-						tc.kvStore(),
-						log.Log,
-						nil, // proxy URL
-						netxlite.NewStdlibResolver(log.Log),
-					)
-					defer netx.Close()
-
-					client := netx.NewHTTPClient()
-					resp, err := client.Get("https://www.example.com/")
-					if err != nil {
-						t.Fatal(err)
-					}
-					defer resp.Body.Close()
-					if resp.StatusCode != tc.expectStatus {
-						t.Fatal("StatusCode: expected", tc.expectStatus, "got", resp.StatusCode)
-					}
-					data, err := netxlite.ReadAllContext(context.Background(), resp.Body)
-					if err != nil {
-						t.Fatal(err)
-					}
-					if diff := cmp.Diff(tc.expectBody, data); diff != "" {
-						t.Fatal(diff)
-					}
-				})
-			})
 		}
 	})
 }
